fix(ci): unblock quality-gate and smoke for tri-state auth PR
- Extract CommunityAdminCard out of CommunityAdmin.tsx so the file drops
  back under the 600-line budget (the new `ready`/AccessDenied guard had
  pushed it to 626, newly tripping the quality gate). Also dedup the page
  skeleton into a local PageSkeleton.
- Update the homepage smoke assertions to the post-#1598 hero copy
  ("Karma helps funders fund and track…") and the "How Karma works"
  section; the old "funding software that does the work" heading and FAQ
  moved off the homepage, so the smoke test was failing on main.
